requirements.txt
Flask
Jinja2
Werkzeug
Flask-Script
Flask-SQLAlchemy
Flask-WTF
Flask-Mail
py-bcrypt
WeasyPrint
boto